Move tunnel class to its own file
